chore: move base model to model directory
